Support optimal compression level for Brotli
The original code defaulted to using the fastest compression level. When
trying to pass in the optimal compression level, the Brotli stream
compression would result in 0 bytes being generated. This fix is
desribed in the next paragraph.

When CompressionLevel.Optimal is being used, and in general, the
destination compression stream should be flushed before trying to
extract the bytes from the underlying stream. This change adds the
necessary calls to FlushAsync().

Duplicate code has also been moved into helper methods
ToCompressedStringAsync() and FromCompressedStringAsync().

Test driver program has been updated to explicitly pass in the
compression level.